Dynamic Treatment Description Page
Added data layer support for treatment description page and allows for data from db to pass through using intents

+++ app/src/main/java/com/example/medikan/ptsd_treatment/DisplayTimerActivity.java
@@ -75,7 +75,7 @@ public class DisplayTimerActivity extends AppCompatActivity {
      * WARNING converting from long to int may mess up some larger numbers. Might be a bug here in the future
      * @param timerValue
      */
-    private void setupTimer(int timerValue) {
+    private void setupTimer(long timerValue) {
 
         final Context context = this;
